# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GachaPull - Pokemon & One Piece Gacha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      summary: Get coin transaction history
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinTransactionListResponse"

  # PAYMENTS
...
  /payments/stripe/create-session:
    post:
      operationId: createStripeSession
      tags: [payments]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "200":
          description: Stripe session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeSessionResponse"
...
  /payments/midtrans/create:
    post:
      operationId: createMidtransPayment
      tags: [payments]
      summary: Create Midtrans payment (QRIS, bank transfer, e-wallet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "200":
          description: Midtrans payment token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MidtransPaymentResponse"
...
  /payments/usdt/create:
    post:
      operationId: createUsdtPayment
      tags: [payments]
      summary: Create USDT payment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "200":
          description: USDT payment address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsdtPaymentResponse"
  /payments/history:
    get:
      operationId: getPaymentHistory
      tags: [payments]
      summary: Get my payment history
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentHistoryResponse"

  # LEADERBOARD
...
components:
  schemas:
...
    CoinTransactionListResponse:
      type: object
      required: [transactions, total, page, limit]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/CoinTransaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CoinTransaction:
      type: object
      required: [id, amount, type, description, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: integer
        type:
          type: string
          enum: [credit, debit]
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    # Payment schemas
...
    CreatePaymentBody:
      type: object
      required: [coinPackageId]
      properties:
        coinPackageId:
          type: integer
    StripeSessionResponse:
      type: object
      required: [sessionId, url]
      properties:
        sessionId:
          type: string
        url:
          type: string
    MidtransPaymentResponse:
      type: object
      required: [token, redirectUrl, orderId]
      properties:
        token:
          type: string
        redirectUrl:
          type: string
        orderId:
          type: string
    UsdtPaymentResponse:
      type: object
      required: [address, amountUsdt, orderId, expiresAt]
      properties:
        address:
          type: string
        amountUsdt:
          type: number
        orderId:
          type: string
        expiresAt:
          type: string
          format: date-time
        qrCodeUrl:
          type: string
          nullable: true
    PaymentHistoryResponse:
      type: object
      required: [payments, total, page, limit]
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PaymentRecord:
      type: object
      required: [id, method, amountUsd, coinsGranted, status, createdAt]
      properties:
        id:
          type: integer
        method:
          type: string
          enum: [stripe, midtrans, usdt]
        amountUsd:
          type: number
        coinsGranted:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time

    # Leaderboard
...